requirements.txt
python-pptx
pillow
pytesseract
sentence-transformers
faiss-cpu
fastapi
uvicorn[standard]
groq
python-dotenv
